fix: move command to get server domains by UUID and remove non existing domain endpoint
