--- cmd/domains/list.go
@@ -1,41 +1,0 @@
-package domains
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/coollabsio/coolify-cli/internal/cli"
-	"github.com/coollabsio/coolify-cli/internal/output"
-	"github.com/coollabsio/coolify-cli/internal/service"
-	"github.com/spf13/cobra"
-)
-
-// NewListCommand creates the list command
-func NewListCommand() *cobra.Command {
-	return &cobra.Command{
-		Use:   "list",
-		Short: "List all domains",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			ctx := context.Background()
-
-			client, err := cli.GetAPIClient(cmd)
-			if err != nil {
-				return fmt.Errorf("failed to get API client: %w", err)
-			}
-
-			domainSvc := service.NewDomainService(client)
-			domains, err := domainSvc.List(ctx)
-			if err != nil {
-				return fmt.Errorf("failed to list domains: %w", err)
-			}
-
-			format, _ := cmd.Flags().GetString("format")
-			formatter, err := output.NewFormatter(format, output.Options{})
-			if err != nil {
-				return err
-			}
-
-			return formatter.Format(domains)
-		},
-	}
-}
